refactor(EmojiPicker): migrate type definitions to a separate file and enhance type safety
