+++ src/components/molecules/EmojiPicker/index.ts
@@ -1,1 +1,2 @@
 export * from './EmojiPicker';
+export * from './EmojiPicker.types';
